refactor(apiplatform): modernize admin UI, add dependencies and CSRF protection
- Add strict_types, copyright headers and #[\Override] to all admin
  Block/Controller/Model/Helper/Setup files
- Move _setForcedFormKeyActions to preDispatch in Role/User controllers
- Migrate Varien_Db_* references to Maho\Db\* in install script
- Add API Platform and Symfony dependencies to composer.json
- Add PSR-4 autoload entry for Maho\ApiPlatform namespace
- Cast storeId to int in rest.php

+++ public/rest.php
@@ -80,7 +80,7 @@ if (strpos($requestUri, '/api/admin/') !== false) {
     if ($isAuthenticated && $adminUser) {
         $_SERVER['MAHO_ADMIN_USER_ID'] = $adminUser->getId();
         $_SERVER['MAHO_ADMIN_USERNAME'] = $adminUser->getUsername();
-        $_SERVER['MAHO_STORE_ID'] = $input['variables']['storeId'] ?? 1;
+        $_SERVER['MAHO_STORE_ID'] = (int) ($input['variables']['storeId'] ?? 1);
     }
 }
